pyproject.toml
[project]
name = "office-party-game"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.5.2",
    "pillow>=11.2.1",
    "python-dotenv>=1.1.0",
    "reflex>=0.7.14",
    "streamlit>=1.45.1",
]
